Refactor and add more opa tests, including robustness tests.
