+++ magda-registry-api/src/test/resources/data/add-esri-access-control-aspect-empty.json
@@ -1,0 +1,42 @@
+[
+    {
+        "id": "record-0",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    },
+    {
+        "id": "record-1",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    },
+
+    {
+        "id": "record-2",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    },
+
+    {
+        "id": "record-3",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    },
+
+    {
+        "id": "record-4",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    },
+
+    {
+        "id": "record-5",
+        "aspects": {
+            "esri-access-control": {}
+        }
+    }
+]
