+++ magda-registry-api/src/test/resources/data/add-esri-access-control-aspect-missing-expiration.json
@@ -1,0 +1,64 @@
+[
+    {
+        "id": "record-0",
+        "aspects": {
+            "esri-access-control": {
+                "groups": ["Dep. A"],
+                "owner": "user0"
+            }
+        }
+    },
+    {
+        "id": "record-1",
+        "aspects": {
+            "esri-access-control": {
+                "groups": ["Dep. A", "Branch A, Dep. A"],
+                "owner": "user1"
+            }
+        }
+    },
+
+    {
+        "id": "record-2",
+        "aspects": {
+            "esri-access-control": {
+                "groups": ["Dep. A", "Branch B, Dep. A"],
+                "owner": "user2"
+            }
+        }
+    },
+
+    {
+        "id": "record-3",
+        "aspects": {
+            "esri-access-control": {
+                "groups": [
+                    "Dep. A",
+                    "Branch B, Dep. A",
+                    "Section C, Branch B, Dep. A"
+                ],
+                "owner": "user3"
+            }
+        }
+    },
+
+    {
+        "id": "record-4",
+        "aspects": {
+            "esri-access-control": {
+                "access": "public",
+                "owner": "user3"
+            }
+        }
+    },
+
+    {
+        "id": "record-5",
+        "aspects": {
+            "esri-access-control": {
+                "groups": ["Dep. A", "Branch B, Dep. A"],
+                "owner": "user2"
+            }
+        }
+    }
+]
