+++ magda-registry-api/src/test/resources/data/add-esri-access-control-aspect-only-has-expiration.json
@@ -1,0 +1,54 @@
+[
+    {
+        "id": "record-0",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    },
+    {
+        "id": "record-1",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    },
+
+    {
+        "id": "record-2",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    },
+
+    {
+        "id": "record-3",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    },
+
+    {
+        "id": "record-4",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    },
+
+    {
+        "id": "record-5",
+        "aspects": {
+            "esri-access-control": {
+                "expiration": 9569380434535
+            }
+        }
+    }
+]
