+++ magda-registry-api/src/test/scala/au/csiro/data61/magda/opa/TestRecordEsriOpaPolicyWithExpirationOnly.scala
@@ -3,14 +3,14 @@ package au.csiro.data61.magda.opa
 import org.scalatest.Ignore
 
 @Ignore
-class TestRecordEsriPolicyWithGroupsOnly
-    extends RecordOpaPolicyWithEsirGroupsOrMagdaOrgUnitsOnlySpec
+class TestRecordEsriOpaPolicyWithExpirationOnly
+    extends RecordEsriOpaPolicyWithInvalidAccessControlAspectSpec
     with GroupsJwtForRecordEsriOpaPolicyTest {
 
   override def beforeAll(): Unit = {
     super.beforeAll()
     testRecords = getTestRecords(
-      dataPath + "add-esri-access-control-aspect-groups-only.json"
+      dataPath + "add-esri-access-control-aspect-only-has-expiration.json"
     )
   }
